backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pandas==2.1.3
scikit-learn>=1.5.0
numpy>=1.26,<2.0
matchms>=0.24.0
gensim>=4.4.0
spec2vec>=0.9.1
python-multipart==0.0.6
aiofiles==23.2.1
qdrant-client==1.11.3
llama-index-core>=0.10.0
llama-index-vector-stores-qdrant>=0.2.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-ollama>=0.1.0
rdkit>=2023.9.1
pdfminer.six>=20221105
python-docx>=1.1.0
